Refactor Schema data type.
This commit contains the changes described in https://github.com/restatedev/restate/issues/3303

* metadata contains the "implementation" of the schema registry, and the serializable data structure `Schema`.
* `metadata/mod.rs` contains the definition of `Schema` and the implementation of the various schema registry APIs. All the fields are private, and access should be available only through schema registry APIs or `SchemaUpdater`.
* `Schema` now uses the new data types from the old `v2.rs` file. On deserialization, we infer the "active service revisions" index (the logic for that is within `serde_hacks`).
* `v2.rs` conversions was moved to `metadata/serde_hacks.rs`, the new data model is directly in `metadata/mod.rs`. Inside it, I moved the old data types *Schemas, which should not be used anymore.
* `SchemaUpdater` was moved inside `metadata`, and deals with the new `Schema` data structure. The behavior remains untouched.
* Switch default writing using the new schema data structure

Everything else is more or less copying stuff around and renaming.
